% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxPredict.R
\name{fluxPredict}
\alias{fluxPredict}
\title{Make predictions from seahorse data by sampling and minimizing total flux}
\usage{
fluxPredict(model, seahorse_data, nsamples = 150)
}
\arguments{
\item{model}{An object of class \code{\link[sybil]{modelorg}}}

\item{seahorse_data}{A data.frame}

\item{nsamples}{The number of samples to perform}
}
\description{
This function integrates the seahorse measurements as constraints into the specified model,
minimizes total flux for each sample, and returns a matrix of reactions x samples where the entries are predicted fluxes.
}
\author{
Alfred Ramirez <akram@bu.edu>
}

