% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxvar_rewrite.R
\name{fluxvar_rewrite}
\alias{fluxvar_rewrite}
\title{Rewrite fluxVar result into a dataframe}
\usage{
fluxvar_rewrite(o)
}
\arguments{
\item{model}{A object of class \code{\link[sybil]{modelorg}}}

\item{reactions}{A character vector of reaction ids}

\item{solver}{The solver to use.  Default SYBIL_SETTINGS("SOLVER")}
}
\description{
This function rewrites an object of \code{\link[sybil]{optsol}}
}
\author{
Alfred Ramirez <akram@bu.edu>
}

