% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seahorse_data.R
\docType{data}
\name{seahorse_data}
\alias{seahorse_data}
\title{Seahorse XF Analyzer data on brown and white adipocytes
Measures of OCR and PPR during the oxdiative stress test on brown and white culture adipocytes.}
\format{A data.frame}
\usage{
data(seahorse_data)
}
\description{
Seahorse XF Analyzer data on brown and white adipocytes
Measures of OCR and PPR during the oxdiative stress test on brown and white culture adipocytes.
}
\keyword{datasets}

