\name{sybilXF-package}
\alias{sybilXF-package}
\alias{sybilXF}
\docType{package}
\title{
\packageTitle{sybilXF}
}
\description{
\packageDescription{sybilXF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sybilXF}
\packageIndices{sybilXF}

}
\author{
\packageAuthor{sybilXF}

Maintainer: \packageMaintainer{sybilXF}
}
\references{

}
\keyword{ }
\seealso{
}
\examples{

#Not run
#data(seahorse_data)
#data(Recon_21A_reduced)
#data_summary <- summarize_seahorse(seahorse_data)
#ba_samples <- sample_seahorse(data_summary, nsamples=2, sample="BA")
#ba_samples <- convert_units(ba_samples, mwr=0.026)
#ba_constraints <- map_seahorse(ba_samples)
#ba_predicted_fluxes <- fluxPredict(model = Recon_21A_reduced, seahorse_data= ba_constraints)

}
