\name{sybilXF-package}
\alias{sybilXF-package}
\alias{sybilXF}
\docType{package}
\title{
\packageTitle{sybilXF}
}
\description{
\packageDescription{sybilXF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sybilXF}
\packageIndices{sybilXF}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{sybilXF}

Maintainer: \packageMaintainer{sybilXF}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ 
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
